Refactor widget tool. [skip ci]
