+++ .gitignore
@@ -11,5 +11,9 @@ __pycache__/
 .idea
 .venv
 
+# VS Code
+.vscode/
+
 # Scicat Widget Development Configuration
 SCICAT_WIDGET_DEBUGGING.yml
+6-scicat/notebooks/myProposals/
